stack navigator 구조 수정
